+++ Examples/AdventureWorks2019/Comparison/01-Compare-Database-Subsets.ps1
@@ -17,7 +17,7 @@ $sessionId = Start-SqlSizerSession -Database $database -ConnectionInfo $connecti
 $sessionId2 = Start-SqlSizerSession -Database $database -ConnectionInfo $connection -DatabaseInfo $info
 
 # Define start set
-$query = New-Object -TypeName Query2
+$query = New-Object -TypeName SqlSizerQuery
 $query.State = [TraversalState]::Include  # Use modern TraversalState enum for forward traversal
 $query.Schema = "Person"
 $query.Table = "Person"