+++ Examples/AdventureWorks2019/Removal/03-Data-Removal-Advanced.ps1
@@ -19,14 +19,14 @@ $info = Get-DatabaseInfo -Database $database -ConnectionInfo $connection
 $sessionId = Start-SqlSizerSession -Database $database -ConnectionInfo $connection -DatabaseInfo $info
 
 # Define start set
-$query = New-Object -TypeName Query2
+$query = New-Object -TypeName SqlSizerQuery
 $query.State = [TraversalState]::InboundOnly  # Use modern TraversalState enum for removal/incoming FK traversal
 $query.Schema = "Person"
 $query.Table = "Person"
 $query.KeyColumns = @('BusinessEntityID')
 $query.Where = "[`$table].FirstName = 'Michael'"
 
-$query2 = New-Object -TypeName Query2
+$query2 = New-Object -TypeName SqlSizerQuery
 $query2.State = [TraversalState]::InboundOnly  # Use modern TraversalState enum for removal/incoming FK traversal
 $query2.Schema = "Production"
 $query2.Table = "Product"