+++ Examples/AdventureWorks2019/Removal/06-Iterative-Removal-Alternative-1.ps1
@@ -25,7 +25,7 @@ while ($true)
     $sessionId = Start-SqlSizerSession -Database $database -ConnectionInfo $connection -DatabaseInfo $info -Installation $false -SecureViews $false -ExportViews $false -Removal $true
 
     # Define start set
-    $query = New-Object -TypeName Query2
+    $query = New-Object -TypeName SqlSizerQuery
     $query.State = [TraversalState]::InboundOnly  # Use modern TraversalState enum for removal/incoming FK traversal
     $query.Schema = "Person"
     $query.Table = "PhoneNumberType"